+++ CHANGELOG.rst
@@ -4,7 +4,7 @@ Changelog
 2.4 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Add a method to import a deck from VDB
 
 
 2.3 (2020-12-21)